tasks/external_data/load-transit/requirements.txt
google-cloud-bigquery==3.14.1
functions-framework==3.5.0
